+++ NEWS.md
@@ -7,6 +7,7 @@
 * New standardized getter `loglik()` for class `Learner`.
 * New measures `"aic"` and `"bic"` to compute the Akaike Information Criterion
   or the Bayesian Information Criterion, respectively.
+* New Resampling method: `ResamplingCustomCV`.
 * Measures now have formal hyperparameters. A popular example where this is
   required is the F1 score, now implemented with customizable `beta`.
 * Fixed hashing of some measures.