+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3 0.12.0-9000
+
 # mlr3 0.12.0
 
 * New method to assign labels to columns in tasks: `Task$label()`.